pyproject.toml
[tool.poetry]
name = "buildwithai"
version = "0.1.0"
description = "Add project description here"
authors = ["Larry Dawson <larry.dawson@gmail.com>"]
license = "BSD-3-Clause"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pydantic = "^2.9.0"
jira = "^3.8.0"
openai = "^1.43.1"
python-dotenv = "^1.0.1"
gitpython = "^3.1.43"
langchain = "^0.3.0"
langchain_community = "^0.3.0"
langchain-openai = "^0.2.0"
fire = "^0.6.0"
tenacity = "^8.4.0"
atlassian-python-api = "^3.41.16"

[tool.poetry.group.dev.dependencies]
black = "^24.8.0"
isort = "^5.13.2"
flake8 = "^7.1.1"
mypy = "^1.11.2"
bandit = "^1.7.9"
docformatter = "^1.7.5"
pre-commit = "^3.8.0"
jupyterlab = "^4.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py311', 'py312']
skip-string-normalization = true
exclude = '''
/(
    \.eggs|
    \.git|
    \.hg|
    \.mypy_cache|
    \.tox|
    \.venv|
    _build|
    buck-out|
    build|
    dist
)/
'''

[tool.isort]
profile = "black"
line_length = 88
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
combine_as_imports = true
use_parentheses = true
ensure_newline_before_comments = true
lines_after_imports = 2

[tool.flake8]
max-line-length = 88
ignore = ["W291"]

[tool.mypy]
python_version = 3.8
ignore_missing_imports = true
strict_optional = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
warn_unused_ignores = true
warn_return_any = true
warn_unreachable = true

[tool.bandit]
skips = []
exclude_dirs = ['tests', 'docs']

[tool.docformatter]
line_length = 88
wrap_summaries = 88
wrap_descriptions = 88
exclude = '''
/(
    \.eggs|
    \.git|
    \.hg|
    \.mypy_cache|
    \.tox|
    \.venv|
    _build|
    buck-out|
    build|
    dist
)/
'''

